requirements.txt
openai==1.74.0
python-dotenv==1.1.0
fastapi==0.115.12
starlette==0.46.1
uvicorn==0.34.0
fastapi
PyMuPDF==1.25.5
python-docx
pdf2image
pytesseract
pillow